// resc/WebApplication/src/modules/lists/models/speciality-list/speciality-list-item.dto.ts
import { NomenclatureDto } from "src/modules/nomenclature/common/models/nomenclature-dto";
import { SpecialityMinisterDto } from "./speciality-minister.dto";

export class SpecialityListItemDto {
  id: number | null;
  specialityListId: number;
  speciality: NomenclatureDto;
  researchArea: NomenclatureDto;
  institution: NomenclatureDto;
  educationalQualification: NomenclatureDto;
  educationFormType: NomenclatureDto;
  specialityMinisters: SpecialityMinisterDto[] = [];
  studentsCount: number;

  constructor() {
    this.specialityMinisters = [new SpecialityMinisterDto()];
  }
}